Maintain sufficient allocated space in the vector and support multiple return values
`wasm_runtime_call_wasm()`` is called with `argv`
used to pass parameters, and the callee reuses
`argv` to store the results. Therefore, `argv`
must be long enough to hold the results.

Typically, a vector's length changes with
operations, but in our scenario, `append()`
operation occurs outside of our control. If there
are more results than parameters, the vector will
end up with an incorrect length unless we
proactively manage it.

Update Dockerfile and devcontainer configuration

Co-Author: Sieluna<seele.peng@outlook.com>