fix: use static trampoline for host function test
This test was inconsistently failing for me (presumably why it was ignored) and figured it was due to the function pointer. Using a static trampoline function and registering that seemed to do the trick. There may be more elegant ways to do this but this seemed like the easiest. I also updated the register function on the builder to include an example of this pattern.